requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.2.0
bidict==0.22.1
black==23.1.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
flake8==6.0.0
Flask==2.2.3
Flask-SocketIO==5.3.2
flatbuffers==23.3.3
frozenlist==1.3.3
fsspec==2023.3.0
gast==0.4.0
google-auth==2.16.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.3
h5py==3.8.0
idna==3.4
importlib-metadata==6.0.0
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.11.0
libclang==15.0.6.1
lightning-utilities==0.7.1
Markdown==3.4.1
MarkupSafe==2.1.2
mccabe==0.7.0
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.24.2
oauthlib==3.2.2
opencv-python==4.4.0.40
opt-einsum==3.3.0
packaging==23.0
pathspec==0.11.0
Pillow==9.4.0
platformdirs==3.1.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pyflakes==3.0.1
python-engineio==4.3.4
python-socketio==5.7.2
pytorch-lightning==1.9.4
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.15.0
socketIO-client==0.7.2
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
tomli==2.0.1
torch==1.13.1
torchmetrics==0.11.3
torchvision==0.14.1
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.14
websocket-client==1.5.1
Werkzeug==2.2.3
wrapt==1.15.0
yarl==1.8.2
zipp==3.15.0
